Additional lex/parse bug fixes
Fixed the following bugs: 1. the lexer was allowing the print statement to
contain uppercase characters and digits in strings. 2. the parser was
allowing the print statement to contain multiple digits. 3. the lexer now
processes long lines of chars and digits as IDs and Digits, respectively.
The parser then catches them as invalid. Still need to ensure longest
length matching during lexing process.