Initial semantic analysis commit
After six hours of work last night and six hours today, the untested
basis of my semantic analyzer is complete. Bad practice to not commit
regularly, but I could not get my GitHub working for a while - forced to
reinstall the program. The analyzer is mainly one big case statement
with a handful of helper functions.